data: Drop gnome-terminal related rules
It is already fixed upstream for quite a while. And the current
situation just puts things into the apps.slice which is not optimal but
still acceptable as we don't need to modify the defaults for the
application slice.

i.e. the only effect is that terminal windows will get a higher share of
resources when competing with other applications.