data: Change method of workaround for the lack of memory_recursiveprot
Current systemd (before v247) do not set memory_recursiveprot.
As such, we need to do some extra steps to protect the children of
session.slice.

We used to do this using DisableControllers=memory, but systemd then
refuses to set the memory.low attribute. So switch to using
DefaultMemoryLow= instead and simply set a huge value. This will result
in the correct behaviour as we should only have direct children.