r-pw-monitor: Add audio integration
This serves as a heuristic for detecting interactive applications and
giving them a boost irrespective of whether they are focused or not.

Requirement: libpipewire

Working:
Adds custom pipewire GSource to main loop. It listens to node events
from pipewire-pulse API, creates an RAppInfo and sets the BOOST_AUDIO
flag depending on whether the current application is playing audio.

It calls `r_app_monitor_app_info_changed()` so that r-app-policy
can then act on this changed signal and give an extra
boost to an audio playing application.