r-game-monitor: Add boosting using GameMode
Another area where we can provide additional boost is games.
We use a similar mechanism to boosting applications playing audio,
taking advantage of a background utility called GameMode by Feral
Interactive.

Every time a game is registered or unregistered we receive a signal
from the dbus interface, we then make the app boosted depending on the
respective signal.

Warning in `get_unit_cgroup_path_from_pid` is suppressed as it is
possible we get a unregistered signal after the cgroup directory is
long gone, in this case it's best to return on receiving NULL as path.