pw-monitor: Fix incorrect free of RAppInfo
The application information is owned by the RAppMonitor object and is
kept alive as long as the cgroup exists. The pipewire monitor code has
(as all other code) no business in trying to free it.

Remove the free and also make the function static so it cannot be used
from other files.